Feat/connect with flask

+++ src/main/java/com/sorocode/voice_access_be_demo/member/entity/Member.java
@@ -30,9 +30,9 @@ public class Member extends BaseEntity {
 
     private LocalDate birthday;
 
-    private Short height;
+    private Float height;
 
-    private Short weight;
+    private Float weight;
 
     private String voiceFileLocation;
